requirements.txt
flask==3.0.0
pandas==2.2.0
numpy==1.26.0
plotly==5.19.0
scikit-learn==1.4.0
python-dotenv==1.0.0
groq==1.1.2
